requirements.txt
streamlit==1.31.0
PyMuPDF==1.23.8
pandas==2.2.0
numpy==1.26.3
openai==1.12.0
Unidecode==1.3.7
python-dotenv==1.0.0